+++ lib/zoi.ex
@@ -199,25 +199,6 @@ defmodule Zoi do
   defdelegate boolean(opts \\ []), to: Zoi.Types.Boolean, as: :new
 
   @doc """
-  Defines a datetime type schema.
-
-  ## Example
-
-      iex> schema = Zoi.datetime()
-      iex> Zoi.parse(schema, "2025-08-07T10:04:22+03:00")
-      {:ok, "2025-08-07T10:04:22+03:00"}
-
-      iex> schema = Zoi.datetime()
-      iex> Zoi.parse(schema, 1754646043)
-      {:error, [%Zoi.Error{message: "invalid datetime type", path: []}]}
-
-  Built-in transformations for include:
-      `Zoi.to_datetime/2`
-  """
-  @doc group: "Basic Types"
-  defdelegate datetime(opts \\ []), to: Zoi.Types.Datetime, as: :new
-
-  @doc """
   Defines a schema that accepts any type of input.
 
   This is useful when you want to allow any data type without validation.
@@ -541,8 +522,9 @@ defmodule Zoi do
       {:ok, "550e8400-e29b-41d4-a716-446655440000"}
       iex> Zoi.parse(schema, "invalid-uuid")
       {:error, [%Zoi.Error{message: "invalid UUID format"}]}
-      iex> schema_v8 = Zoi.uuid(version: "v8")
-      iex> Zoi.parse(schema_v8, "6d084cef-a067-8e9e-be6d-7c5aefdfd9b4")
+
+      iex> schema = Zoi.uuid(version: "v8")
+      iex> Zoi.parse(schema, "6d084cef-a067-8e9e-be6d-7c5aefdfd9b4")
       {:ok, "6d084cef-a067-8e9e-be6d-7c5aefdfd9b4"}
   """
   @doc group: "Formats"
@@ -798,19 +780,6 @@ defmodule Zoi do
   def to_downcase(schema) do
     schema
     |> transform({Zoi.Transforms, :transform, [[:to_downcase]]})
-  end
-
-  @doc """
-  Converts a given `Zoi.datetime()` to `DateTime` elixir struct.
-  ## Example
-      iex> schema = Zoi.datetime() |> Zoi.to_datetime()
-      iex> Zoi.parse(schema, "2025-08-07T10:04:22+03:00")
-      {:ok, ~U[2025-08-07 07:04:22Z]}
-  """
-
-  def to_datetime(schema) do
-    schema
-    |> transform({Zoi.Transforms, :transform, [[:to_datetime]]})
   end
 
   @doc """